services/ingestion/requirements.txt
kafka-python-ng==2.1.0
python-dotenv==1.0.0
twitchio==2.7.0
websockets==11.0.3
tensorflow==2.12.0 # Ensure compatibility with your model
numpy==1.23.5 # Pin numpy version for TF